% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_function.r
\name{first_function}
\alias{first_function}
\title{A dog Function}
\usage{
first_function(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE.}
}
\description{
This function allows you to express your love of dogs.
}
\examples{
first_function()
}
\keyword{cats}

